Add collectables
As the player moves around the board they collect things. This required the addition of game state and play mode was refactored to be based on modifying state rather than a board specifically. This has the added benefit that activating play mode doesn't disrupt the board that was created in build mode.