Implement falling rocks
Rocks fall if there is empty space or falling rocks under them. As long as they are falling the player cannot make any further moves so that they don't disrupt the falling rocks.

If a rock lands on another rock and there is empty space to the left or right at the rock's level and below then it will fall in that direction so that falling towers of rocks collapse downwards into piles.