Load boards from an encoded string
This change allows boards to be loaded from an encoded string. Boards are now encoded in a kind of base 64 encoding. To make this encoding readable boards now have a maximum width and height and there is a maximum run length for the run length encoding. Longer maximum run lengths increase the size of the board's encoded form and so they must be balanced with likely run lengths.

At a future time it may be interesting to make the maximum run length depend dynamically on the width of the board to see if that decreases the size of the encoded board.