Evaluate state changes locally
`applyGravityToFallingRocks` attempted to track and handle all falling rocks within the board as a whole. This proved very difficult to track, understand, and extend.

However, it's pretty clear that the behavior of a falling rock only depends on the tiles in its immediate neighborhood. This change switches to explicitly use this method to track and handle falling rocks. As a benefit, it also uses this method to track player state changes.