Initial water implementation
This implements the initial behavior for water. Water flows around objects and presents another obstacle to avoid.

This change is missing significant tests because I feel that the state manager needs a major refactor. Currently all tests need to happen through the single update method and these tests are both difficult to write and to debug.