Simplify dependencies
This cleans up the dev container to use the image directly since the Dockerfile doesn't do anything interesting.

It also removes Mocha and Chai in preference for Node's built in assertion and test runner since they are more than sufficient for a project like this. Node is upgraded to the latest LTS to get stable versions of the APIs.