Convert logic to a state machine
At its core, this program can be run as a state machine. Until now it seemed simpler to write the logic more procedurally but adding water proved difficult to conceptualize because it was difficult to see the various cases and interactions in the procedural logic.

This change moves to a state machine with a pattern matcher and transition function to make the large number of possible states more manageable. It is now much easier to see the various states and confirm the transitions make sense. Debugging this has also been much easier.

More tests are added because the transitions make it clearer what needs to be tested.

Each tile now also tracks if it was updated in the current cycle in order to ensure states that update tiles to their left don't update too quickly.