Add conveyors
This change adds conveyors as a map element. Each tile now has an
additional property specifying which direction, if any, a player will
automatically move if placed on that tile. Conveyors move the player in
any of the cardinal directions.

Only players are moved by conveyors but a player on a conveyor can also
move a rock as if the player were moving on their own.

If a conveyor cannot move a player because there is an obstacle in the
way then the player dies.

If an object is moved into a player, the player also dies.