Add a richer encoding for tiles and boards
When the encoding was originally added the tiles and boards were very
simple. However, as things have developed tiles have gained new features
that are not currently represented in the encoding such as flow
direction.

The new encoding allows all of a tile's properties to be encoded and
decoded. The encoding is longer than the previous base-64 encoding but
is less fragile to new tiles being added or new properties being
introduced.

Signed-off-by: Colin Jeanne <colinjeanne@hotmail.com>