Add richer tile updates
So far tile updates have been raw tiles and this has meant that
additional patterns had to be created to handle propagating properties
from one tile to its update.

This change converts tile updates into a function that takes the
currently matched region so that the function can propagate the
properties that are relevant. As a result, we can reduce the number of
patterns we need to check and create.