Handle player movement with the state machine
This fixes #13

Much of the player input logic was handled outside of the state machine
and duplicated logic handled in other states. This change moves all of
this logic into the state machine.

Living players now also have an input direction which causes them to
move, if possible, in that direction. Once this state is handled the
input direction automatically returns to no movement.

Moving player input handling into the state machine also makes it
possible to allow players to move while the rest of the state machine is
running since their input is handled on the same clock as the rest of
the simulation.

Signed-off-by: Colin Jeanne <colinjeanne@hotmail.com>