Add debug logging for pattern matching
Since moving to a state machine it has been difficult to debug issues
with the patterns. Debugging patterns involves tediously looking at the
pattern and the region in the debugger to figure out what's happening.

This change prints a debug log in the console showing which pattern was
matched and against what tile.