Expand patterns to a larger area
Fixes #31 and fixes #28

The 3x3 patterns are not sufficient to handle all scenarios. For
example, scenarios where players push rocks requires that logical
center of the pattern be offset from the actual center so that there is
enough space to decide whether the rock can move.

This change expands patterns so that they have access to move horizontal
space and so that the logic center of the tile is always at the same
place within the pattern. This allows better control over the precedence
between different patterns because they are now partitioned based on the
primary tile.

As part of this change, rocks now prefer to fall to the right rather
than the left because in these cases the empty tile that the rock falls
into is the primary tile of the pattern and tiles are updated right to
left.