Falling rocks kill and roll off players
Fixes #33

The rule that caused falling rocks to be killed by falling rocks was too
general of a rule for its location in the list of rules. General rules
should come before more specific rules in order to allow the more
specific rules to accurately run.

In this case, the location of this general rule caused me to not
understand that specific rules were necessary to handle the case where a
rock could fall and roll off a player. The rule that causes a rock to
roll to the right actually ends up running before the rule killing the
player because the empty tile to the right of the player is what pulls
the rock into it.

The fix is to add new rules that specially handle rocks rolling off of
living players and to move the more general rule after these more
specific rules.