Players conveyed right are killed by rocks
Fixes #37

When a player is conveyed right into an empty space next to a rock and
then conveyed into the rock the player is killed even if they could move
the rock.

The issue is that when a tile is updated it only updates the tiles that
it is directly adjacent to. However, patterns match tiles in a wider
range. For this case, the tile to the right of the rock was not adjacent
to the player and so not marked for an update but the pattern that moves
the rock is applied centered at that tile.

The fix is to mark all tiles that are in range of a pattern as updated
whenever a tile is changed to ensure all possible patterns have a chance
to match.