Implement doors
This change adds doors and keys. There are four different colors of
doors and keys where a player must be holding a key of the correct color
in order to open a door. Opening a door uses up the key and removes the
door.

A player can only hold one key at a time and walking over another key
does not pick it up if the player is already holding a key.